backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
alembic==1.15.2
pydantic-settings==2.8.1
python-dotenv==1.1.0
transformers==4.51.3
torch==2.7.0
httpx==0.28.1
websockets==15.0.1
pytest==8.3.5
